+++ modules/functions_manager/builtin/cpp/ismacroBuiltin.cpp
@@ -1,0 +1,31 @@
+//=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of the Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// SPDX-License-Identifier: LGPL-3.0-or-later
+// LICENCE_BLOCK_END
+//=============================================================================
+#include "ismacroBuiltin.hpp"
+#include "Error.hpp"
+#include "PathFuncManager.hpp"
+#include "characters_encoding.hpp"
+#include "CheckerHelpers.hpp"
+//=============================================================================
+using namespace Nelson;
+//=============================================================================
+ArrayOfVector
+Nelson::FunctionsGateway::ismacroBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
+{
+    ArrayOfVector retval;
+    nargincheck(argIn, 1, 1);
+    nargoutcheck(nLhs, 0, 1);
+    ArrayOf param1 = argIn[0];
+    std::wstring name = param1.getContentAsWideString();
+    wstringVector filenames;
+    bool res = PathFuncManager::getInstance()->find(name, filenames);
+    retval << ArrayOf::logicalConstructor(res);
+    return retval;
+}
+//=============================================================================